Complete Part 4 Section D and full Part 4

+++ part4/bloglist-backend/controllers/blogs.js
@@ -1,7 +1,7 @@
 const blogsRouter = require('express').Router()
 
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -14,36 +14,48 @@ blogsRouter.get('/', async (request, response) => {
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const body = request.body
+blogsRouter.post(
+  '/',
+  middleware.userExtractor,
+  async (request, response) => {
+    const body = request.body
+    const user = request.user
 
-  const user = await User.findById(body.userId)
+    if (!user) {
+      return response.status(401).json({
+        error: 'token missing or invalid'
+      })
+    }
 
-  if (!user) {
-    return response.status(400).json({
-      error: 'invalid user'
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0,
+      user: user._id
     })
+
+    const savedBlog = await blog.save()
+
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
+    const populatedBlog = await Blog.findById(savedBlog._id)
+      .populate('user', {
+        username: 1,
+        name: 1
+      })
+
+    response.status(201).json(populatedBlog)
   }
-
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
-    user: user._id
-  })
-
-  const savedBlog = await blog.save()
-
-  user.blogs = user.blogs.concat(savedBlog._id)
-
-  await user.save()
-
-  response.status(201).json(savedBlog)
-})
+)
 
 blogsRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
+    .populate('user', {
+      username: 1,
+      name: 1
+    })
 
   if (blog) {
     response.json(blog)
@@ -52,15 +64,41 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
+blogsRouter.delete(
+  '/:id',
+  middleware.userExtractor,
+  async (request, response) => {
+    const user = request.user
 
-  if (deletedBlog) {
+    if (!user) {
+      return response.status(401).json({
+        error: 'token missing or invalid'
+      })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+      return response.status(404).end()
+    }
+
+    if (blog.user.toString() !== user.id.toString()) {
+      return response.status(403).json({
+        error: 'only creator can delete a blog'
+      })
+    }
+
+    await Blog.findByIdAndDelete(request.params.id)
+
+    user.blogs = user.blogs.filter(
+      blogId => blogId.toString() !== request.params.id
+    )
+
+    await user.save()
+
     response.status(204).end()
-  } else {
-    response.status(404).end()
   }
-})
+)
 
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
@@ -72,7 +110,7 @@ blogsRouter.put('/:id', async (request, response) => {
     likes: body.likes
   }
 
-  const blog = await Blog.findByIdAndUpdate(
+  const updated = await Blog.findByIdAndUpdate(
     request.params.id,
     updatedBlog,
     {
@@ -81,8 +119,8 @@ blogsRouter.put('/:id', async (request, response) => {
     }
   )
 
-  if (blog) {
-    response.json(blog)
+  if (updated) {
+    response.json(updated)
   } else {
     response.status(404).end()
   }
